# query/physicalplan: Use arrow compute package for binary scalar expressions