--- a/query/physicalplan/binaryscalarexpr.go
+++ b/query/physicalplan/binaryscalarexpr.go
@@ -2,11 +2,13 @@ package physicalplan
 
 import (
 	"bytes"
+	"context"
 	"errors"
 	"fmt"
 
 	"github.com/apache/arrow/go/v14/arrow"
 	"github.com/apache/arrow/go/v14/arrow/array"
+	"github.com/apache/arrow/go/v14/arrow/compute"
 	"github.com/apache/arrow/go/v14/arrow/scalar"
 
 	"github.com/polarsignals/frostdb/query/logicalplan"
@@ -79,111 +81,8 @@ func (e BinaryScalarExpr) String() string {
 var ErrUnsupportedBinaryOperation = errors.New("unsupported binary operation")
 
 func BinaryScalarOperation(left arrow.Array, right scalar.Scalar, operator logicalplan.Op) (*Bitmap, error) {
+	// TODO: Figure out dictionary arrays and lists with compute next
 	leftType := left.DataType()
-	switch leftType {
-	case arrow.FixedWidthTypes.Boolean:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return BooleanArrayScalarNotEqual(left.(*array.Boolean), right.(*scalar.Boolean))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case &arrow.FixedSizeBinaryType{ByteWidth: 16}:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return FixedSizeBinaryArrayScalarNotEqual(left.(*array.FixedSizeBinary), right.(*scalar.FixedSizeBinary))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case arrow.BinaryTypes.String:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return StringArrayScalarNotEqual(left.(*array.String), right.(*scalar.String))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case arrow.BinaryTypes.Binary:
-		switch operator {
-		case logicalplan.OpEq:
-			switch r := right.(type) {
-			case *scalar.Binary:
-				return ArrayScalarEqual(left, r)
-			case *scalar.String:
-				return ArrayScalarEqual(left, r.Binary)
-			default:
-				panic("something terrible has happened, this should have errored previously during validation")
-			}
-		case logicalplan.OpNotEq:
-			switch r := right.(type) {
-			case *scalar.Binary:
-				return BinaryArrayScalarNotEqual(left.(*array.Binary), r)
-			case *scalar.String:
-				return BinaryArrayScalarNotEqual(left.(*array.Binary), r.Binary)
-			default:
-				panic("something terrible has happened, this should have errored previously during validation")
-			}
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case arrow.PrimitiveTypes.Int64:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return Int64ArrayScalarNotEqual(left.(*array.Int64), right.(*scalar.Int64))
-		case logicalplan.OpLt:
-			return Int64ArrayScalarLessThan(left.(*array.Int64), right.(*scalar.Int64))
-		case logicalplan.OpLtEq:
-			return Int64ArrayScalarLessThanOrEqual(left.(*array.Int64), right.(*scalar.Int64))
-		case logicalplan.OpGt:
-			return Int64ArrayScalarGreaterThan(left.(*array.Int64), right.(*scalar.Int64))
-		case logicalplan.OpGtEq:
-			return Int64ArrayScalarGreaterThanOrEqual(left.(*array.Int64), right.(*scalar.Int64))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case arrow.PrimitiveTypes.Int32:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return Int32ArrayScalarNotEqual(left.(*array.Int32), right.(*scalar.Int32))
-		case logicalplan.OpLt:
-			return Int32ArrayScalarLessThan(left.(*array.Int32), right.(*scalar.Int32))
-		case logicalplan.OpLtEq:
-			return Int32ArrayScalarLessThanOrEqual(left.(*array.Int32), right.(*scalar.Int32))
-		case logicalplan.OpGt:
-			return Int32ArrayScalarGreaterThan(left.(*array.Int32), right.(*scalar.Int32))
-		case logicalplan.OpGtEq:
-			return Int32ArrayScalarGreaterThanOrEqual(left.(*array.Int32), right.(*scalar.Int32))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	case arrow.PrimitiveTypes.Uint64:
-		switch operator {
-		case logicalplan.OpEq:
-			return ArrayScalarEqual(left, right)
-		case logicalplan.OpNotEq:
-			return Uint64ArrayScalarNotEqual(left.(*array.Uint64), right.(*scalar.Uint64))
-		case logicalplan.OpLt:
-			return Uint64ArrayScalarLessThan(left.(*array.Uint64), right.(*scalar.Uint64))
-		case logicalplan.OpLtEq:
-			return Uint64ArrayScalarLessThanOrEqual(left.(*array.Uint64), right.(*scalar.Uint64))
-		case logicalplan.OpGt:
-			return Uint64ArrayScalarGreaterThan(left.(*array.Uint64), right.(*scalar.Uint64))
-		case logicalplan.OpGtEq:
-			return Uint64ArrayScalarGreaterThanOrEqual(left.(*array.Uint64), right.(*scalar.Uint64))
-		default:
-			panic("something terrible has happened, this should have errored previously during validation")
-		}
-	}
-
 	switch arr := left.(type) {
 	case *array.Dictionary:
 		switch operator {
@@ -201,7 +100,50 @@ func BinaryScalarOperation(left arrow.Array, right scalar.Scalar, operator logic
 		panic("TODO: list comparisons unimplemented")
 	}
 
+	switch operator {
+	case logicalplan.OpEq:
+		return ArrayScalarCompute("equal", left, right)
+	case logicalplan.OpNotEq:
+		return ArrayScalarCompute("not_equal", left, right)
+	case logicalplan.OpLt:
+		return ArrayScalarCompute("less", left, right)
+	case logicalplan.OpLtEq:
+		return ArrayScalarCompute("less_equal", left, right)
+	case logicalplan.OpGt:
+		return ArrayScalarCompute("greater", left, right)
+	case logicalplan.OpGtEq:
+		return ArrayScalarCompute("greater_equal", left, right)
+	}
+
 	return nil, ErrUnsupportedBinaryOperation
+}
+
+func ArrayScalarCompute(funcName string, left arrow.Array, right scalar.Scalar) (*Bitmap, error) {
+	equalsResult, err := compute.CallFunction(context.TODO(), funcName, nil, compute.NewDatum(left), compute.NewDatum(right))
+	if err != nil {
+		return nil, fmt.Errorf("error calling equal function: %w", err)
+	}
+	defer equalsResult.Release()
+	equalsDatum, ok := equalsResult.(*compute.ArrayDatum)
+	if !ok {
+		return nil, fmt.Errorf("expected *compute.ArrayDatum, got %T", equalsResult)
+	}
+	equalsArray, ok := equalsDatum.MakeArray().(*array.Boolean)
+	if !ok {
+		return nil, fmt.Errorf("expected *array.Boolean, got %T", equalsDatum.MakeArray())
+	}
+	defer equalsArray.Release()
+
+	res := NewBitmap()
+	for i := 0; i < equalsArray.Len(); i++ {
+		if equalsArray.IsNull(i) {
+			continue
+		}
+		if equalsArray.Value(i) {
+			res.AddInt(i)
+		}
+	}
+	return res, nil
 }
 
 func DictionaryArrayScalarNotEqual(left *array.Dictionary, right scalar.Scalar) (*Bitmap, error) {
@@ -278,369 +220,6 @@ func DictionaryArrayScalarEqual(left *array.Dictionary, right scalar.Scalar) (*B
 			if dict.Value(left.GetValueIndex(i)) == string(data) {
 				res.Add(uint32(i))
 			}
-		}
-	}
-
-	return res, nil
-}
-
-func FixedSizeBinaryArrayScalarNotEqual(left *array.FixedSizeBinary, right *scalar.FixedSizeBinary) (*Bitmap, error) {
-	res := NewBitmap()
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if !bytes.Equal(left.Value(i), right.Data()) {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func StringArrayScalarNotEqual(left *array.String, right *scalar.String) (*Bitmap, error) {
-	res := NewBitmap()
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if left.Value(i) != string(right.Data()) {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func BinaryArrayScalarNotEqual(left *array.Binary, right *scalar.Binary) (*Bitmap, error) {
-	res := NewBitmap()
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if !bytes.Equal(left.Value(i), right.Data()) {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int64ArrayScalarNotEqual(left *array.Int64, right *scalar.Int64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if left.Value(i) != right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int64ArrayScalarLessThan(left *array.Int64, right *scalar.Int64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) < right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int64ArrayScalarLessThanOrEqual(left *array.Int64, right *scalar.Int64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) <= right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int64ArrayScalarGreaterThan(left *array.Int64, right *scalar.Int64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) > right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int64ArrayScalarGreaterThanOrEqual(left *array.Int64, right *scalar.Int64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) >= right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func ArrayScalarEqual(left arrow.Array, right scalar.Scalar) (*Bitmap, error) {
-	res := NewBitmap()
-	switch arr := left.(type) {
-	case *array.Boolean:
-		rightType, ok := right.(*scalar.Boolean)
-		if !ok {
-			return nil, fmt.Errorf("expected scalar.Boolean, got %T", right)
-		}
-		for i := 0; i < arr.Len(); i++ {
-			if arr.IsNull(i) {
-				continue
-			}
-			if arr.Value(i) == rightType.Value {
-				res.Add(uint32(i))
-			}
-		}
-	case *array.FixedSizeBinary:
-		rightType, ok := right.(*scalar.FixedSizeBinary)
-		if !ok {
-			return nil, fmt.Errorf("expected scalar.FixedSizeBinary, got %T", right)
-		}
-		for i := 0; i < arr.Len(); i++ {
-			if arr.IsNull(i) {
-				continue
-			}
-			if bytes.Equal(arr.Value(i), rightType.Data()) {
-				res.Add(uint32(i))
-			}
-		}
-	case *array.Int32:
-		rightType, ok := right.(*scalar.Int32)
-		if !ok {
-			return nil, fmt.Errorf("expected scalar.Int32, got %T", right)
-		}
-		for i := 0; i < arr.Len(); i++ {
-			if arr.IsNull(i) {
-				continue
-			}
-			if arr.Value(i) == rightType.Value {
-				res.Add(uint32(i))
-			}
-		}
-	case *array.Int64:
-		rightType, ok := right.(*scalar.Int64)
-		if !ok {
-			return nil, fmt.Errorf("expected scalar.Int64, got %T", right)
-		}
-		for i := 0; i < arr.Len(); i++ {
-			if arr.IsNull(i) {
-				continue
-			}
-			if arr.Value(i) == rightType.Value {
-				res.Add(uint32(i))
-			}
-		}
-	case *array.Uint64:
-		rightType, ok := right.(*scalar.Uint64)
-		if !ok {
-			return nil, fmt.Errorf("expected scalar.Uint64, got %T", right)
-		}
-		for i := 0; i < arr.Len(); i++ {
-			if arr.IsNull(i) {
-				continue
-			}
-			if arr.Value(i) == rightType.Value {
-				res.Add(uint32(i))
-			}
-		}
-	default:
-		return nil, fmt.Errorf("unsupported array type: %T", arr)
-	}
-
-	return res, nil
-}
-
-func Uint64ArrayScalarNotEqual(left *array.Uint64, right *scalar.Uint64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if left.Value(i) != right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Uint64ArrayScalarLessThan(left *array.Uint64, right *scalar.Uint64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) < right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Uint64ArrayScalarLessThanOrEqual(left *array.Uint64, right *scalar.Uint64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) <= right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Uint64ArrayScalarGreaterThan(left *array.Uint64, right *scalar.Uint64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) > right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Uint64ArrayScalarGreaterThanOrEqual(left *array.Uint64, right *scalar.Uint64) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) >= right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func BooleanArrayScalarNotEqual(left *array.Boolean, right *scalar.Boolean) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) != right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int32ArrayScalarNotEqual(left *array.Int32, right *scalar.Int32) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			res.Add(uint32(i))
-			continue
-		}
-		if left.Value(i) != right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int32ArrayScalarLessThan(left *array.Int32, right *scalar.Int32) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) < right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int32ArrayScalarLessThanOrEqual(left *array.Int32, right *scalar.Int32) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) <= right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int32ArrayScalarGreaterThan(left *array.Int32, right *scalar.Int32) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) > right.Value {
-			res.Add(uint32(i))
-		}
-	}
-
-	return res, nil
-}
-
-func Int32ArrayScalarGreaterThanOrEqual(left *array.Int32, right *scalar.Int32) (*Bitmap, error) {
-	res := NewBitmap()
-
-	for i := 0; i < left.Len(); i++ {
-		if left.IsNull(i) {
-			continue
-		}
-		if left.Value(i) >= right.Value {
-			res.Add(uint32(i))
 		}
 	}