--- a/query/physicalplan/binaryscalarexpr_test.go
+++ b/query/physicalplan/binaryscalarexpr_test.go
@@ -12,7 +12,7 @@ import (
 
 func BenchmarkBinaryScalarOperation(b *testing.B) {
 	ab := array.NewInt64Builder(memory.DefaultAllocator)
-	for i := int64(0); i < 10_000_000; i++ {
+	for i := int64(0); i < 1_000_000; i++ {
 		ab.Append(i % 10)
 	}
 
@@ -21,7 +21,20 @@ func BenchmarkBinaryScalarOperation(b *testing.B) {
 
 	s := scalar.NewInt64Scalar(4) // chosen by fair dice roll. guaranteed to be random.
 
-	for i := 0; i < b.N; i++ {
-		_, _ = BinaryScalarOperation(arr, s, logicalplan.OpEq)
+	operators := []logicalplan.Op{
+		logicalplan.OpEq,
+		logicalplan.OpNotEq,
+		logicalplan.OpLt,
+		logicalplan.OpLtEq,
+		logicalplan.OpGt,
+		logicalplan.OpGtEq,
+	}
+
+	for _, op := range operators {
+		b.Run(op.String(), func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				_, _ = BinaryScalarOperation(arr, s, op)
+			}
+		})
 	}
 }