Instead of implementing individual functions for each BinaryScalar operation, we'll now use the arrow `compute` package. It already implements most of the heavy lifting for us. 

Once we pass in the array and the scalar, we can convert the data back to a `*array.Boolean`. Walking this array, we construct the bitmap like we did before. 
It cleans up a lot of duplicate code and removes the need to add new functions for new Arrow types. 

Given the quality of live improvement, the performance impact is fine:
```
name                         old time/op    new time/op    delta
BinaryScalarOperation/==-24    3.39ms ± 2%    3.76ms ± 7%  +11.00%  (p=0.008 n=5+5)
BinaryScalarOperation/!=-24    11.2ms ± 3%    11.4ms ± 1%   +1.64%  (p=0.032 n=5+4)
BinaryScalarOperation/<-24     6.00ms ± 6%    6.25ms ± 3%     ~     (p=0.056 n=5+5)
BinaryScalarOperation/<=-24    7.07ms ± 2%    7.04ms ± 1%     ~     (p=0.841 n=5+5)
BinaryScalarOperation/>-24     7.02ms ± 6%    6.92ms ± 5%     ~     (p=0.690 n=5+5)
BinaryScalarOperation/>=-24    7.78ms ± 3%    7.95ms ± 5%     ~     (p=0.310 n=5+5)

name                         old alloc/op   new alloc/op   delta
BinaryScalarOperation/==-24     509kB ± 0%     643kB ± 0%  +26.23%  (p=0.008 n=5+5)
BinaryScalarOperation/!=-24     532kB ± 0%     665kB ± 0%  +25.12%  (p=0.008 n=5+5)
BinaryScalarOperation/<-24      532kB ± 0%     532kB ± 0%     ~     (all equal)
BinaryScalarOperation/<=-24     532kB ± 0%     532kB ± 0%     ~     (all equal)
BinaryScalarOperation/>-24      532kB ± 0%     532kB ± 0%     ~     (all equal)
BinaryScalarOperation/>=-24     532kB ± 0%     532kB ± 0%     ~     (all equal)

name                         old allocs/op  new allocs/op  delta
BinaryScalarOperation/==-24       263 ± 0%       309 ± 0%  +17.49%  (p=0.008 n=5+5)
BinaryScalarOperation/!=-24       267 ± 0%       313 ± 0%  +17.23%  (p=0.008 n=5+5)
BinaryScalarOperation/<-24        267 ± 0%       267 ± 0%     ~     (all equal)
BinaryScalarOperation/<=-24       267 ± 0%       267 ± 0%     ~     (all equal)
BinaryScalarOperation/>-24        267 ± 0%       267 ± 0%     ~     (all equal)
BinaryScalarOperation/>=-24       267 ± 0%       267 ± 0%     ~     (all equal)
```