@asubiotto 2024-01-11
Interesting. I guess you can even get rid of the type switch in this case and just delegate that to `compute.CallFunction`? I wonder what else in our execution engine we could replace with these compute kernels.