@metalmatze 2024-01-11
I'm working on an update with these exact changes. It's going to delete _a lot_ of code. :joy: