@gernest 2024-01-11
@metalmatze @asubiotto @thorfour  I take this opportunity to thank you for being awesome and giving me a chance to hack on `frostdb` uninterrupted.

I have a minor surgery on Monday that will take me a couple of weeks to recover. So my silence won't mean I jumped ship, but rather I won't be able to work on my laptop.

I will pickup the slack the moment I get back, I'm quite experienced with `compute` and whole `arrow/**`  stuff, so I will probably focus there and our parquet converter ro see how much improvement we can get.

I have some pending local branches, that I will be glad to share the details with anyone touching these topics maybe they can be helpful.

- Building direct arrow records in logictest
- Integrating `GenericTable` in `parca` Ingester
- Benchmarking `arrowutils.ReorderRecord`  


 shameless plug 😃 **_Please don't forget about about me If you ever need to expand your team, our collaboration has been awesome._**