@metalmatze 2024-01-11
@gernest, thank you for all your contributions in the last weeks. They've been totally awesome!
Good luck with the surgery! I wish you a fast recovery! 

We'll definitely let you know if we hire someone (you'll probably see it quickly anyway).